refactor(mobile): drep id modal and push notifications request
